refactor(aggressive.h, behavior_enum.h, behaviors.h, braitenberg_vehicle, coward, explore, factory_bv, factory_food, graphics_arena_viewer,love.h): refactor #4 <Adding strategy pattern for behaviors>
- Got rid of behavior_enum for low coupling
- Adding aggressive, coward, none, explore, love, and behaviors class
for strategy pattern. Each have a different GetWheelVelocity function
that applies to each behavior
- GetWheelVelocity function takes in the depencies to the wheel velocity
and calculates the correct velocity for eacfh behavior. Then it changes
that type using a pointer to the wheel_velocity.
- For graphics Arena Viewer, got rid of enumerated behavior types
therefore added a getBehaviorType and getIndexOfBehavior to know the
type of behavior
- Factory Unit test for BV have to fix for when getting food_behavior
you get what the behavior type is which we converted to a string instead
of the enum type